requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
reportlab==4.2.5
qrcode[pil]==8.0
Pillow==10.4.0
alembic==1.13.3
python-dotenv==1.0.1
markdown==3.7
bleach==6.1.0
